amd/compiler: fix fix_exports() with p_discard_if
A second s_endpgm can appear when p_discard_if is used.
Previously, this would insert a null export if there was not a export
between the beginning of the basic block and the last s_endpgm.

This changes it so that it will insert a null export if there was not a
export between:
- the beginning of the basic block or the second-to-last s_endpgm (if any)
- and the last s_endpgm.