aco: Use the LLVM disassembler API to print assembly
The current solution of creating a temp file and calling llvm-mc is a
complete hack. Using the API means no one has to patch the driver to
point to their LLVM, plus the finer-grained API means we can do a much
better job pinpointing invalid instructions and recovering afterwards.
Now the precise invalid words are printed, and we can skip past the
entire word instead of a single byte like LLVM does, leading to a much
better chance we get to the intended sequence of instructions. In the
future we can also print labels for basic blocks and have LLVM print
them instead of a raw offset for branches.