amd/compiler: implement proper WQM handling
This adds a new pseudo-instruction: p_wqm. Usage of this instruction
indicates that the operand should be computed for the whole quad.
Also, MUBUF and MIMG instructions have a new field (set during instruction
selection) "disable_wqm" which indicates that the instruction should
not be executed for helper invocations.

If a shader doesn't have any p_wqm instructions or doesn't set
disable_wqm, lower_wqm() is either a no-op or simple inserts a s_wqm_b64
at the beginning of the shader.