amd/compiler: ensure the exp_cnt for a export's operands is 0
map::emplace() doesn't modify the map when the key already exists.

This change is so that correct s_waitcnt instructions will be inserted for
code like this:
exp %1122:v[1], %1122:v[1], undef, undef
exp %1122:v[1], %1122:v[1], undef, undef
 v1: %1105:v[1] = v_cvt_pkrtz_f16_f32 %1083:v[4], %1098:s[0]

Previously a s_waitcnt with an export count of 1 would be inserted,
instead of 0.